+++ lib/compo-thasauce-fetcher.js
