+++ lib/party.js
@@ -7,12 +7,12 @@ const mm = require('music-metadata');
 const nodeshout = require('nodeshout-napi');
 const { MessageEmbed } = require('discord.js');
 
-const CompoThaSauceFetcher = require('./compo_thasauce_fetcher');
+const CompoThaSauceFetcher = require('./compo-thasauce-fetcher');
 const { metadataUpdater } = require('./machines');
 const Song = require('./song');
-const RoundFetcher = require('./round_fetcher');
-const RoundTranscoder = require('./round_transcoder');
-const RoundAnnouncer = require('./round_announcer');
+const RoundFetcher = require('./round-fetcher');
+const RoundTranscoder = require('./round-transcoder');
+const RoundAnnouncer = require('./round-announcer');
 const { announcerFinal, transcodeFinal } = require('../utils/symbols');
 const fetchEnv = require('../utils/fetch-env');
