+++ lib/round-transcoder.js
